Reorganized file structure. Code seems to be operating normally, but time will tell. We have still to write the testing suite.
